Start of adding Authenticated API routes
All api routes are now authenticated with Cognito User Pool
to ensure that only authenticated users can access the API.
Using the Amplify library, the user is authenticated and the
JWT token is passed to the API. The API then validates the
token and returns the data.

The api requires a Bearer token to be passed in the header
of the request.